+++ kernel/Panic.hpp
